refactor: car 패키지 내부로 car 관련 클래스 이동
