+++ src/main/java/racingcar/car/Car.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.car;
 
 import racingcar.utils.RandomUtils;
