+++ src/main/java/racingcar/car/CarInputService.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.car;
 
 import racingcar.utils.IOUtils;
