+++ src/main/java/racingcar/car/CarInputValidator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.car;
 
 import racingcar.constant.ErrorMessage;
